# feat(bindings/python): read APIs return `memoryview` instead of `bytes` to avoid copy