--- a/.github/workflows/bindings_python.yml
+++ b/.github/workflows/bindings_python.yml
@@ -52,12 +52,28 @@ jobs:
         working-directory: "bindings/python"
         run: |
           python -m pip install -e .[test]
-      - name: Run behave
+      - name: Run pytest
         working-directory: "bindings/python"
         env:
           OPENDAL_MEMORY_TEST: on
         run: |
           pytest -vk TestMemory
+
+  sdist:
+    runs-on: ubuntu-latest
+    if: "startsWith(github.ref, 'refs/tags/')"
+    steps:
+      - uses: actions/checkout@v4
+      - uses: PyO3/maturin-action@v1
+        with:
+          working-directory: "bindings/python"
+          command: sdist
+          args: -o dist
+      - name: Upload sdist
+        uses: actions/upload-artifact@v3
+        with:
+          name: wheels
+          path: bindings/python/dist
 
   linux:
     runs-on: ubuntu-latest
@@ -69,7 +85,7 @@ jobs:
           manylinux: auto
           working-directory: "bindings/python"
           command: build
-          args: --release --sdist -o dist
+          args: --release -o dist --find-interpreter
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with:
@@ -85,7 +101,7 @@ jobs:
         with:
           working-directory: "bindings/python"
           command: build
-          args: --release -o dist
+          args: --release -o dist --find-interpreter
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with:
@@ -102,7 +118,7 @@ jobs:
           working-directory: "bindings/python"
           command: build
           target: universal2-apple-darwin
-          args: --release -o dist
+          args: --release -o dist --find-interpreter
       - name: Upload wheels
         uses: actions/upload-artifact@v3
         with: