--- a/bindings/python/Cargo.toml
+++ b/bindings/python/Cargo.toml
@@ -34,6 +34,6 @@ doc = false
 [dependencies]
 futures = "0.3.28"
 opendal.workspace = true
-pyo3 = { version = "0.19", features = ["abi3-py37"] }
+pyo3 = "0.19"
 pyo3-asyncio = { version = "0.19", features = ["tokio-runtime"] }
 tokio = "1"