--- a/bindings/python/src/asyncio.rs
+++ b/bindings/python/src/asyncio.rs
@@ -27,9 +27,11 @@ use pyo3::exceptions::PyIOError;
 use pyo3::exceptions::PyNotImplementedError;
 use pyo3::exceptions::PyStopAsyncIteration;
 use pyo3::exceptions::PyValueError;
+use pyo3::ffi;
 use pyo3::prelude::*;
 use pyo3::types::PyBytes;
 use pyo3::types::PyDict;
+use pyo3::AsPyPointer;
 use pyo3_asyncio::tokio::future_into_py;
 use tokio::io::AsyncReadExt;
 use tokio::io::AsyncSeekExt;
@@ -39,6 +41,7 @@ use crate::build_operator;
 use crate::build_opwrite;
 use crate::format_pyerr;
 use crate::layers;
+use crate::Buffer;
 use crate::Entry;
 use crate::Metadata;
 use crate::PresignedRequest;
@@ -74,8 +77,15 @@ impl AsyncOperator {
         let this = self.0.clone();
         future_into_py(py, async move {
             let res: Vec<u8> = this.read(&path).await.map_err(format_pyerr)?;
-            let pybytes: PyObject = Python::with_gil(|py| PyBytes::new(py, &res).into());
-            Ok(pybytes)
+            Python::with_gil(|py| {
+                let buffer = Buffer::from(res).into_py(py);
+                unsafe {
+                    PyObject::from_owned_ptr_or_err(
+                        py,
+                        ffi::PyMemoryView_FromObject(buffer.as_ptr()),
+                    )
+                }
+            })
         })
     }
 
@@ -330,8 +340,15 @@ impl AsyncReader {
                     buffer
                 }
             };
-            let pybytes: PyObject = Python::with_gil(|py| PyBytes::new(py, &buffer).into());
-            Ok(pybytes)
+            Python::with_gil(|py| {
+                let buffer = Buffer::from(buffer).into_py(py);
+                unsafe {
+                    PyObject::from_owned_ptr_or_err(
+                        py,
+                        ffi::PyMemoryView_FromObject(buffer.as_ptr()),
+                    )
+                }
+            })
         })
     }