… bindings/python/python/opendal/__init__.pyi