Before

```python
In [1]: import opendal

In [2]: op = opendal.Operator("memory")

In [3]: op.write("test", bytes(10 * 1024 * 1024))

In [4]: %timeit op.read("test")
627 µs ± 3.9 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

After

```python
In [1]: import opendal

In [2]: op = opendal.Operator("memory")

In [3]: op.write("test", bytes(10 * 1024 * 1024))

In [4]: %timeit op.read("test")
291 µs ± 761 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

Unfortunately better performance comes with a price, we're no longer able to build `abi3` wheels which means we need to build and publish one wheel per Python version, OS and CPU architecture.

We can remove the `unsafe` once https://github.com/PyO3/pyo3/pull/3514 is released.